+++ R/get_lists.R
@@ -7,6 +7,8 @@
 #'
 #' @return Data frame with columns `FullName`, `Id`, `Name`, and `TaxonomyId`.
 #'
+#' @seealso [list_groups()], [list_datasets()]
+#'
 #' @export
 #'
 #' @examples
@@ -31,6 +33,7 @@ list_species <-
 #' Get list of available groups
 #'
 #' @param species Optional species name, for just the groups for that species.
+#' @param group Optional name for a specific group
 #' @param url URL for GeneNetwork API
 #'
 #' @return Data frame with columns `FullName`, `Id`, `Name`, and `TaxonomyId`.
@@ -42,10 +45,21 @@ list_species <-
 #' @examples
 #' list_groups()
 #' list_groups("barley")
+#' list_groups("rat", "HSNIH-Palmer")
+#' list_groups(group="HSNIH-Palmer")
 list_groups <-
-    function(species=NULL, url=gnapi_url())
+    function(species=NULL, group=NULL, url=gnapi_url())
 {
-    if(!is.null(species)) {
+    if(!is.null(group) && group != "") { # note this uses group/ rather than groups/
+        query <- "group/"
+        if(!is.null(species) && species != "") {
+            query <- paste0(query, "/", species)
+        }
+        result <- query_gn(paste0(query, "/", group), url=url)
+        if(is.null(result)) return(NULL)
+        return(as.data.frame(result))
+    }
+    else if(!is.null(species) && species != "") {
         stopifnot(length(species)==1)
         result <- query_gn(paste0("groups/", species), url=url)
     } else {
@@ -61,6 +75,7 @@ list_groups <-
 #' List available datasets for a group.
 #'
 #' @param group Name of group, as single character string
+#' @param dataset Optional name of a specific dataset
 #' @param url URL for GeneNetwork API
 #'
 #' @return A data frame with dataset ID, name, and description
@@ -72,10 +87,17 @@ list_groups <-
 #' @examples
 #' list_datasets("BXD")
 list_datasets <-
-    function(group, url=gnapi_url())
+    function(group, dataset=NULL, url=gnapi_url())
 {
     # group should be a single character string
     stopifnot(length(group) == 1)
+
+    if(!is.null(dataset) && dataset != "") {
+        result <- query_gn(paste0("dataset/", group, "/", dataset),
+                           url=url)
+        if(is.null(result)) return(NULL)
+        return(as.data.frame(result))
+    }
 
     result <- query_gn(paste0("datasets/", group), url=url)